Moves rules to `meta_rules`
