+++ src/meta_rules/lines.rs
@@ -87,7 +87,7 @@ impl Lines {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use range::Range;
     use std::rc::Rc;
