+++ src/meta_rules/number.rs
@@ -68,7 +68,7 @@ impl Number {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use range::Range;
     use std::rc::Rc;
