+++ src/meta_rules/optional.rs
@@ -53,7 +53,7 @@ impl Optional {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use range::Range;
     use std::rc::Rc;
