+++ src/meta_rules/repeat.rs
@@ -61,7 +61,7 @@ impl Repeat {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use std::rc::Rc;
     use range::Range;
