+++ src/meta_rules/select.rs
@@ -53,7 +53,7 @@ impl Select {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use range::Range;
     use std::rc::Rc;
