+++ src/meta_rules/separated_by.rs
@@ -83,7 +83,7 @@ impl SeparatedBy {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use std::rc::Rc;
     use range::Range;
