+++ src/meta_rules/token.rs
@@ -61,7 +61,7 @@ impl Token {
 
 #[cfg(test)]
 mod tests {
-    use super::super::*;
+    use all::*;
     use std::rc::Rc;
     use range::Range;
